dom/base/test/test_bug811701.xhtml

<html xmlns="http://www.w3.org/1999/xhtml">
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=811701
-->
<head>
  <title>Test for Bug 811701</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=811701">Mozilla Bug 811701</a>
<p id="display"></p>
<div id="content" style="display: none">
<math xmlns="http://www.w3.org/1998/Math/MathML"><mtext>test</mtext></math>
<svg xmlns="http://www.w3.org/2000/svg"><polygon points="0,0 100,100 200,300"/></svg>
</div>
<pre id="test">
</pre>
  <script type="application/javascript">
  <![CDATA[


  /** Test for Bug 811701 **/
  var math = document.querySelector("math");
  is(math.innerHTML,
     '<mtext xmlns="http://www.w3.org/1998/Math/MathML">test</mtext>',
     "<math> should have innerHTML");
  is(math.outerHTML,
     '<math xmlns="http://www.w3.org/1998/Math/MathML"><mtext>test</mtext></math>',
     "<math> should have innerHTML");
  math.innerHTML = "<mo>+</mo>";
  is(math.firstChild.namespaceURI, "http://www.w3.org/1998/Math/MathML",
     "Should have the right namespace after setting innerHTML on <math>");

  var polygon = document.querySelector("polygon");
  is(polygon.parentNode.innerHTML,
     '<polygon xmlns="http://www.w3.org/2000/svg" points="0,0 100,100 200,300"/>',
     "<svg> should have innerHTML");
  is(polygon.parentNode.outerHTML,
     '<svg xmlns="http://www.w3.org/2000/svg"><polygon points="0,0 100,100 200,300"/></svg>',
     "<svg> should have outerHTML");
  is(polygon.outerHTML, '<polygon xmlns="http://www.w3.org/2000/svg" points="0,0 100,100 200,300"/>',
     "<polygon> should have outerHTML");

  var svg = document.querySelector("svg");
  svg.innerHTML = "<rect/>";
  is(svg.firstChild.namespaceURI, "http://www.w3.org/2000/svg",
     "Should have the right namespace after setting innerHTML on <math>");
  ]]>
</script>
</body>
</html>
